--- backend-danny/.env
@@ -1,4 +1,0 @@
-AZURE_ENDPOINT=https://eastus.api.cognitive.microsoft.com
-AZURE_KEY=a18f0abb556641db9cf284cea4e8e84f
-
-GOOGLE_API_KEY=AlzaSyCX8LNYz8kfbwbGV6h8U29781gz1BUBhuQ
