+++ backend-danny/.gitignore
@@ -1,1 +1,4 @@
 venv/
+
+# .env file
+.env
